Move GdbOsProcessHandler to impl.
